move TestingApp to types mod
