+++ cosmwasm/contracts/yield-ward/src/tests/lpt.rs
@@ -7,13 +7,12 @@ use crate::tests::utils::call::{call_add_token, call_update_token_config};
 use crate::tests::utils::init::{
     get_lp_contract_address_from_response, get_tokens_info,
     instantiate_yield_ward_contract_with_tokens, instantiate_yield_ward_contract_without_tokens,
-    TestingApp,
 };
 use crate::tests::utils::query::{
     get_all_tokens_configs, get_token_config, get_token_denom_by_lpt_address,
     get_token_denom_by_source,
 };
-use crate::tests::utils::types::{TestInfo, TokenTestInfo};
+use crate::tests::utils::types::{TestInfo, TestingApp, TokenTestInfo};
 use lp_token::msg::QueryMsg as lp_token_query_msg;
 
 fn assert_token_config(
