+++ cosmwasm/contracts/yield-ward/src/tests/unstake.rs
@@ -2,11 +2,11 @@ use crate::state::{QueueParams, StakeItem, StakeStatsItem, UnstakeItem};
 use crate::tests::utils::call::{
     call_stake, call_stake_and_unstake, call_stake_response, call_unstake, call_unstake_response,
 };
-use crate::tests::utils::init::{instantiate_yield_ward_contract_with_tokens, TestingApp};
+use crate::tests::utils::init::instantiate_yield_ward_contract_with_tokens;
 use crate::tests::utils::query::{
     get_stake_item, get_stake_params, get_stake_stats, get_unstake_item, get_unstake_params,
 };
-use crate::tests::utils::types::{TestInfo, TokenTestInfo};
+use crate::tests::utils::types::{TestInfo, TestingApp, TokenTestInfo};
 use crate::types::{StakeActionStage, Status, UnstakeActionStage};
 use cosmwasm_std::{Uint128, Uint256};
